add salaryCap helpers. Refactor Team and Trade dashboard into components. Only return players currently assigned to teams from the server
