+++ client/src/Components/TradeSimulator/index.js
@@ -1,13 +1,10 @@
-// @flow
-
 import _ from 'lodash'
 import React, { Component } from 'react'
 import PlayerSelect from '../PlayerSelect'
 import TopNav from '../TopNav'
 import Loading from '../Loading'
-import { Bar } from 'react-chartjs-2'
-import { colors, warnColors } from '../gradients'
-import 'chartjs-plugin-annotation'
+import Chart from './Chart'
+import { calcSalaryCap } from '../../helpers'
 
 export default class TradeSimulator extends Component {
   constructor (props) {
@@ -37,18 +34,16 @@ export default class TradeSimulator extends Component {
   }
 
   applyTrade () {
-    let { playerA, playerB, stats, trades } = this.state
-    let teamA = stats.forPlayer(playerA)['Team']
-    let teamB = stats.forPlayer(playerB)['Team']
+    const { players, trades, playerA, playerB } = this.state
 
-    // stats.applyTrade(playerA, playerB)
-    // applyTrade (playerA: string, playerB: string) {
-    //   let teamA = this.data[playerA]['team']
-    //   let teamB = this.data[playerB]['team']
-    //
-    //   this.data[playerA]['team'] = teamB
-    //   this.data[playerB]['team'] = teamA
-    // }
+    const playerAIdx = _.findIndex(players, (p) => p.name === playerA)
+    const playerBIdx = _.findIndex(players, (p) => p.name === playerB)
+
+    const teamA = players[playerAIdx]['team']
+    const teamB = players[playerBIdx]['team']
+
+    players[playerAIdx]['team'] = teamB
+    players[playerBIdx]['team'] = teamA
 
     trades.push({
       playerA: {
@@ -61,18 +56,26 @@ export default class TradeSimulator extends Component {
       }
     })
 
-    this.setState({stats: stats, trades: trades})
+    this.setState({players: players, trades: trades})
   }
 
-  deleteTrade (trade: Trade) {
-    let { stats, trades } = this.state
-    let playerA = trade.playerA.name
-    let playerB = trade.playerB.name
+  deleteTrade (trade) {
+    const { players, trades } = this.state
+    const playerA = trade.playerA.name
+    const playerB = trade.playerB.name
 
-    stats.applyTrade(playerB, playerA)
+    const playerAIdx = _.findIndex(players, (p) => p.name === playerA)
+    const playerBIdx = _.findIndex(players, (p) => p.name === playerB)
+
+    const teamA = players[playerAIdx]['team']
+    const teamB = players[playerBIdx]['team']
+
+    players[playerAIdx]['team'] = teamB
+    players[playerBIdx]['team'] = teamA
+
     _.remove(trades, (t) => _.isEqual(t, trade))
 
-    this.setState({stats: stats, trades: trades})
+    this.setState({players: players, trades: trades})
   }
 
   renderTrades (trades: Array<any>) {
@@ -103,82 +106,11 @@ export default class TradeSimulator extends Component {
     })
   }
 
-  renderGraph () {
-    const players = this.state.players;
-    const teamNames = _.uniq(players.map(p => p.team));
-    const salaryCap = _.sum(_.map(players, (p) => p.salary)) / 8 * 1.01
-
-    const data = {
-      labels: teamNames,
-      datasets: _.flatten(teamNames.map(team => {
-        const teamPlayers = _.sortBy(players.filter((p) => p.team === team), (p) => p.salary)
-        return teamPlayers.map((player, idx) => {
-          const teamSalary = _.sum(_.map(teamPlayers, (p) => p.salary))
-          const overCap = teamSalary > salaryCap
-
-          return {
-            type: 'bar',
-            label: player.name,
-            stack: team,
-            data: [player.salary],
-            backgroundColor: overCap ? warnColors[idx] : colors[idx],
-            hoverBackgroundColor: overCap ? warnColors[idx] : colors[idx]
-          }
-        })
-      }))
-    }
-
-    const options = {
-      legend: {
-        display: false
-      },
-      tooltips: {
-        callbacks: {
-          title: (tooltipItem, data) => {
-            const item = data.datasets[tooltipItem[0].datasetIndex]
-            return item.stack
-          }
-        }
-      },
-      scales: {
-        xAxes: [{
-          barPercentage: 0.6,
-          categoryPercentage: 1.0,
-          ticks: {
-            autoSkip: false
-          }
-        }],
-        yAxes: [{
-          stacked: true
-        }]
-      },
-      animation: {
-        duration: 0
-      },
-      annotation: {
-        annotations: [{
-          type: 'line',
-          mode: 'horizontal',
-          scaleID: 'y-axis-0',
-          value: salaryCap,
-          borderColor: 'black',
-          borderWidth: 2,
-          label: {
-            position: 'right',
-            backgroundColor: 'black',
-            content: 'Salary Cap',
-            enabled: true
-          }
-        }]
-      }
-    }
-
-    return <Bar data={data} redraw={true} options={options}/>
-  }
-
   renderMain () {
     const { players, playerA, playerB, trades } = this.state
+    const teamNames = _.uniq(players.map(p => p.team));
     const playerNames = players.map(p => p.name)
+    const salaryCap = calcSalaryCap(players);
 
     return (
       <div>
@@ -212,7 +144,11 @@ export default class TradeSimulator extends Component {
         </div>
 
         <div className="row" style={{paddingTop: 10}}>
-          { this.renderGraph() }
+          <Chart
+            players={players}
+            teamNames={teamNames}
+            salaryCap={salaryCap}
+          />
         </div>
       </div>
     )
